fix SymbolicRing coercion error
In `_2x2_matrix_entries`, we take a square root involving beta. If beta is an element of SR, then the square root is in SR. When we go to assign the resulting element to a matrix element, we get a coercion error. Instead, map the element into self._ring first, and then take the square root, which will work especially for finite field extensions, but for any ring in which a square root is defined.